Move AbstractEquality* classes to SAGESharp.Testing namespace
