+++ SAGESharpTests/SAGESharpTests.csproj
@@ -78,8 +78,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="AbstractEqualityByRefTests.cs" />
-    <Compile Include="AbstractEqualityByValTests.cs" />
     <Compile Include="MockingExtensions.cs" />
     <Compile Include="SLB\ASCIIExtensionsTests.cs" />
     <Compile Include="SLB\EqualityUtilsTests.cs" />
@@ -108,6 +106,8 @@
     <Compile Include="SLB\UIntExtensionsTests.cs" />
     <Compile Include="Testing\Matchers.cs" />
     <Compile Include="Testing\ParameterGroup.cs" />
+    <Compile Include="Testing\AbstractEqualityByRefTests.cs" />
+    <Compile Include="Testing\AbstractEqualityByValTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
